Moved async result methods to extenion method class. Updates to unit tests.
